+++ classes/Request.php
@@ -2,22 +2,22 @@
 
 namespace Neat\Http;
 
-use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Message\RequestInterface;
 
 /**
  * HTTP Request
  */
 class Request extends Message
 {
-    /** @var ServerRequestInterface */
+    /** @var RequestInterface */
     protected $message;
 
     /**
      * Request constructor
      *
-     * @param ServerRequestInterface $request
+     * @param RequestInterface $request
      */
-    public function __construct(ServerRequestInterface $request)
+    public function __construct(RequestInterface $request)
     {
         parent::__construct($request);
     }
@@ -33,9 +33,9 @@ class Request extends Message
     }
 
     /**
-     * @return ServerRequestInterface
+     * @return RequestInterface
      */
-    public function psr(): ServerRequestInterface
+    public function psr()
     {
         return $this->message;
     }
@@ -77,124 +77,10 @@ class Request extends Message
     }
 
     /**
-     * Get query (aka GET) parameter(s)
-     *
-     * @param string $var
-     * @return mixed
-     */
-    public function query($var = null)
-    {
-        if ($var === null) {
-            return $this->message->getQueryParams();
-        }
-
-        return $this->message->getQueryParams()[$var] ?? null;
-    }
-
-    /**
-     * Get parsed body (aka POST) parameter(s)
-     *
-     * @param string $var
-     * @return mixed
-     */
-    public function post($var = null)
-    {
-        if ($var === null) {
-            return $this->message->getParsedBody();
-        }
-
-        return $this->message->getParsedBody()[$var] ?? null;
-    }
-
-    /**
-     * Get multipart files
-     *
-     * @param array $key
-     * @return null|Upload|Upload[]|Upload[][]|...
-     */
-    public function files(...$key)
-    {
-        $files = $this->toFiles($this->message->getUploadedFiles());
-        while ($key) {
-            $files = $files[array_shift($key)] ?? null;
-        }
-
-        return $files;
-    }
-
-    /**
-     * @param array $files
-     * @return Upload|Upload[]|Upload[][]
-     */
-    protected function toFiles(array $files): array
-    {
-        return array_map(function ($file) {
-            if (is_array($file)) {
-                return $this->toFiles($file);
-            }
-
-            return new Upload($file);
-        }, $files);
-    }
-
-    /**
-     * Get cookie parameter(s)
-     *
-     * @param string $name
-     * @return mixed
-     */
-    public function cookie($name = null)
-    {
-        if ($name === null) {
-            return $this->message->getCookieParams();
-        }
-
-        return $this->message->getCookieParams()[$name] ?? null;
-    }
-
-    /**
-     * @param string $name
-     * @return mixed
-     */
-    public function server(string $name = null)
-    {
-        if ($name === null) {
-            return $this->message->getServerParams();
-        }
-
-        return $this->message->getServerParams()[$name] ?? null;
-    }
-
-    /**
-     * @return string|null
-     */
-    public function clientIp()
-    {
-        return $this->server('REMOTE_ADDR');
-    }
-
-    /**
-     * Set cookie parameters
-     *
-     * @param string $name
-     * @param string $value
-     */
-    protected function setCookie($name, $value = null)
-    {
-        if ($value !== null) {
-            $this->message = $this->message->withCookieParams(array_merge($this->cookie(), [$name => $value]));
-        } elseif ($this->cookie($name)) {
-            $cookies = $this->cookie();
-            unset($cookies[$name]);
-            $this->message = $this->message->withCookieParams($cookies);
-        }
-    }
-
-    /**
      * Get new request with method
      *
      * @param string $method
-     * @return Request
+     * @return static
      */
     public function withMethod($method)
     {
@@ -208,60 +94,12 @@ class Request extends Message
      * Get new request with URL
      *
      * @param Url $url
-     * @return Request
+     * @return static
      */
     public function withUrl(Url $url)
     {
         $new = clone $this;
         $new->message = $this->message->withUri($url->getUri());
-
-        return $new;
-    }
-
-    /**
-     * Get new request with query parameters
-     *
-     * @param array $query
-     * @return Request
-     */
-    public function withQuery(array $query)
-    {
-        $new = clone $this;
-        $new->message = $this->message->withUri($this->message->getUri()->withQuery(http_build_query($query)));
-
-        return $new;
-    }
-
-    /**
-     * Get new request with uploaded files
-     *
-     * @param Upload[]|Upload[][]|... $files
-     * @return Request
-     */
-    public function withFiles(array $files)
-    {
-        array_walk_recursive($files, function (&$upload) {
-            /** @var Upload $upload */
-            $upload = $upload->file();
-        });
-
-        $new          = clone $this;
-        $new->message = $this->message->withUploadedFiles($files);
-
-        return $new;
-    }
-
-    /**
-     * Get new request with cookie parameter
-     *
-     * @param string $name
-     * @param string $value
-     * @return Request
-     */
-    public function withCookie($name, $value)
-    {
-        $new = clone $this;
-        $new->setCookie($name, $value);
 
         return $new;
     }