+++ changelog.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Neat HTTP client and server suggestion.
+- ServerRequest class.
+
+### Changed
+- Request is now only a client-side request.
 
 ### Deprecated
 - Input, Router, Session and Download/Redirect implementations.